Backend/requirements.txt
amqp==5.1.1
aniso8601==9.0.1
async-timeout==4.0.2
bcrypt==3.2.2
billiard==3.6.4.0
blinker==1.5
celery==5.2.7
cffi==1.15.1
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cycler==0.11.0
Deprecated==1.2.13
dnspython==2.2.1
email-validator==1.2.1
Flask==2.0.3
Flask-Cache==0.13.1
Flask-Cors==3.0.10
Flask-Login==0.5.0
Flask-Principal==0.4.0
Flask-RESTful==0.3.9
Flask-Security-Too==4.1.5
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
fonttools==4.30.0
greenlet==1.1.2
idna==3.3
itsdangerous==2.1.0
jedi==0.18.1
Jinja2==3.0.3
kiwisolver==1.4.0
kombu==5.2.4
MarkupSafe==2.1.0
matplotlib==3.5.1
numpy==1.22.3
packaging==21.3
parso==0.8.3
passlib==1.7.4
Pillow==9.0.1
prompt-toolkit==3.0.30
pycparser==2.21
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2022.2.1
redis==4.3.4
six==1.16.0
SQLAlchemy==1.4.31
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.0.3
wrapt==1.14.1
WTForms==3.0.1
